requirements.txt
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
haishoku==1.1.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==7.0.0
psycopg2-binary==2.8.4
requests==2.22.0
SQLAlchemy==1.3.10
urllib3==1.25.8
Werkzeug==0.16.0
WTForms==2.2.1
